DynLoco Documentation

DynLoco is a Julia package for simulating and optimizing simple dynamic locomotion models. The package includes a set of structs that describe models, functions to manipulate models, and optimization routines to perform dynamic optimization on them. The optimization is based on JuMP.jl.

There is presently one main locomotion model, WalkRW2l.

Getting started

The following demonstrates the generation of a simple walking model, continous-time simulation of a single step, and then simulation of multiple discrete steps.

using DynLoco, Plots
w = WalkRW2l() # a default walking model
(t, theta, angvel) = simulatestep(w)
plot(t, theta, xlabel="time", ylabel="theta")
nsteps = multistep(w, Ps = fill(w.P, (10,1)))
plotvees(nsteps, xlabel="step #", ylabel="speed")

DynLoco.WalkRW2l — Type

w = WalkRW2l(vm, P, α, γ, L, M, g, parms, limitcycle)

Rimless wheel in 2D, linearized. Struct containing model parameters. Parameter values can be overridden with WalkRW2l(w, α=0.35) which produces a new model like w except with a new value for alpha. Similarly any parameter may be replaced using named keyword arguments. State: mid-stance speed vm. Control: Push-off impulse P.

parms is a list of numerical model parameters. limitcycle is a named tuple with f, a function whose root is a fixed point. safety (logical) prevents model from falling backwards, rather it always moves forward, albeit very slowly if given insufficient forward speed.

The default object is a limit cycle with speed of 0.4.